Normalized the data & added a line graph
Normalized the data to better detect yearly trends.
And made a new line graph function that plots the normalized data.
Also added comments to all the functions and explained the grouping choices.

Known issues i still need to fix:
The labels of the graphs all show the collumn name, a custom name would be clearer